Rename Jakarta OSGi bridge to jakarta.whiteboard; fix DS activation race in its itest
The Jakarta bridge implements the OSGi Servlet Whiteboard, not a Jakarta
port of the classic HttpService, so name it accordingly and free up
javax.whiteboard for a possible future javax-based Whiteboard bridge:

- org.jupnp.transport.jakarta.httpservice -> org.jupnp.transport.jakarta.whiteboard
  (package org.jupnp.transport.jakarta.httpservice -> org.jupnp.transport.jakarta.whiteboard)
- JakartaHttpServiceServletContainerAdapter -> JakartaWhiteboardServletContainerAdapter
- JakartaHttpServiceSharedStreamServerProvider -> JakartaWhiteboardSharedStreamServerProvider
- itest module org.jupnp.osgi.tests.jetty12.jakarta -> org.jupnp.osgi.tests.jetty12.jakarta.whiteboard,
  matching what it actually exercises (Jetty 12 + the Jakarta Whiteboard bridge)

Updates every cross-reference: bundles/pom.xml and bom/jupnp/pom.xml
module/dependency lists (re-sorted), itests/pom.xml, root pom.xml's
Central-publishing excludeArtifacts (which still had the itest's old
name), and regenerates -runbundles via -DwithResolver.

Also fixes the DS lifecycle race behind the intermittent Java 17 CI
failure in that itest's RealSharedServerActivationTest: enabling
OSGiUpnpServiceConfiguration before the Jakarta Whiteboard bridge's
SharedStreamServerProvider was available let the component activate
without it, then reactivate once the provider showed up --
UpnpServiceImpl.unsetUpnpServiceConfiguration() briefly nulls the
configuration mid-rebind, which a UpnpService reference obtained just
before that reactivation could observe as null. Waiting for
SharedStreamServerProvider first means activation always happens with
the provider already bound, so no later reactivation. Verified with
repeated itest runs (previously intermittent, now consistently green).
