+++ bundles/org.jupnp.transport.jakarta.servlet/pom.xml
@@ -29,7 +29,7 @@
     <!--
       jakarta.servlet-based counterpart of org.jupnp.transport.javax.servlet (which stays on javax.servlet):
       the two namespaces are source-incompatible, so this is a separate package/bundle rather than a
-      shared one, for org.jupnp.transport.jakarta.httpservice's bridge to the OSGi HTTP Whiteboard
+      shared one, for org.jupnp.transport.jakarta.whiteboard's bridge to the OSGi HTTP Whiteboard
       Service for Jakarta Servlet (Compendium R8, org.osgi.service.servlet). Compile scope: this
       bundle's whole purpose is servlet support, so it should propagate to consumers.
     -->
