+++ itests/org.jupnp.osgi.tests.jetty12.jakarta.whiteboard/pom.xml
@@ -1,0 +1,150 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.jupnp.pom</groupId>
+    <artifactId>itests</artifactId>
+    <version>4.0.0-SNAPSHOT</version>
+  </parent>
+
+  <groupId>org.jupnp.itests</groupId>
+  <artifactId>org.jupnp.osgi.tests.jetty12.jakarta.whiteboard</artifactId>
+
+  <name>jUPnP Integration Tests OSGi Jetty 12 Jakarta Whiteboard</name>
+
+  <properties>
+    <basedirRoot>../..</basedirRoot>
+    <maven.compiler.release>17</maven.compiler.release>
+    <!-- Pax Web 11.x: the jakarta.servlet-based line, unlike bom/runtime's javax-based pax.web.version
+         (8.0.25) used by org.jupnp.osgi.tests: the two can't share a property/BOM entry since they're
+         different major versions of the same groupId:artifactId coordinates. Matches the version
+         openhab-core#5479 (Jetty 12/Jakarta EE 10 migration) settled on. -->
+    <pax.web.jakarta.version>11.1.3</pax.web.jakarta.version>
+    <pax.logging.jetty12.version>2.3.5</pax.logging.jetty12.version>
+  </properties>
+
+  <dependencies>
+    <dependency>
+      <groupId>org.jupnp</groupId>
+      <artifactId>org.jupnp</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.jupnp</groupId>
+      <artifactId>org.jupnp.transport.jetty12</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <!--
+      The bundle under test: bridges OSGiUpnpServiceConfiguration to Pax Web's Jakarta HTTP Whiteboard
+      instead of jetty12's standalone StreamServer.
+    -->
+    <dependency>
+      <groupId>org.jupnp</groupId>
+      <artifactId>org.jupnp.transport.jakarta.whiteboard</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <!--
+      Real Jakarta HTTP Whiteboard implementation, running on Jetty 12's ee10 (Jakarta EE 10) servlet
+      environment. This is what proves org.jupnp.transport.jakarta.whiteboard actually works against a
+      real Whiteboard, not just a synthetic HttpServiceRuntime stand-in.
+    -->
+    <dependency>
+      <groupId>org.ops4j.pax.web</groupId>
+      <artifactId>pax-web-jetty</artifactId>
+      <version>${pax.web.jakarta.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.ops4j.pax.web</groupId>
+      <artifactId>pax-web-runtime</artifactId>
+      <version>${pax.web.jakarta.version}</version>
+    </dependency>
+    <!--
+      The bundle that actually tracks OSGi-service-registered Servlets and wires them into the runtime
+      (per Karaf's own pax-web-whiteboard feature = pax-web-core + this bundle). Without it,
+      JakartaWhiteboardServletContainerAdapter.registerServlet()'s context.registerService(Servlet.class,
+      ...) call publishes a service nobody consumes.
+    -->
+    <dependency>
+      <groupId>org.ops4j.pax.web</groupId>
+      <artifactId>pax-web-extender-whiteboard</artifactId>
+      <version>${pax.web.jakarta.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.ops4j.pax.web</groupId>
+      <artifactId>pax-web-spi</artifactId>
+      <version>${pax.web.jakarta.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.ops4j.pax.web</groupId>
+      <artifactId>pax-web-api</artifactId>
+      <version>${pax.web.jakarta.version}</version>
+    </dependency>
+    <!-- pax-web-runtime imports org.apache.tomcat.util.descriptor.web unconditionally (web.xml descriptor
+         parsing), even when the actual HTTP backend is pax-web-jetty, not Tomcat. -->
+    <dependency>
+      <groupId>org.ops4j.pax.web</groupId>
+      <artifactId>pax-web-tomcat-common</artifactId>
+      <version>${pax.web.jakarta.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.jetty.ee10</groupId>
+      <artifactId>jetty-ee10-servlet</artifactId>
+      <version>${jetty12.version}</version>
+    </dependency>
+    <!--
+      Needed as an actual runtime candidate in the resolver's repo index, not just for compilation: the
+      transitive dependency org.jupnp.transport.jakarta.whiteboard/.servlet pull in from bom.compile is
+      compile-scope only and doesn't make the jar available to bnd's -standalone self index.
+    -->
+    <dependency>
+      <groupId>jakarta.servlet</groupId>
+      <artifactId>jakarta.servlet-api</artifactId>
+      <version>${jakarta.servlet.api.version}</version>
+    </dependency>
+    <!--
+      Fragment bundle Pax Web itself publishes to attach the osgi.contract=JakartaServlet capability onto
+      the plain jakarta.servlet-api jar above, which doesn't declare it: pax-web-runtime hard-requires
+      that capability to resolve at all. Confirmed against Karaf's own pax-web-specs feature
+      (assemblies/apache-karaf/.../pax-web-features-*.xml in the Karaf source tree), which pairs the two
+      the same way.
+    -->
+    <dependency>
+      <groupId>org.ops4j.pax.web</groupId>
+      <artifactId>pax-web-compatibility-servlet</artifactId>
+      <version>${pax.web.jakarta.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.osgi</groupId>
+      <artifactId>org.osgi.service.servlet</artifactId>
+      <version>${osgi.service.servlet.version}</version>
+    </dependency>
+    <!--
+      Keep this as a direct dependency so bnd resolver output for this Java 17 module stays stable
+      when runbundles are regenerated with -DwithResolver.
+    -->
+    <dependency>
+      <groupId>org.ops4j.pax.logging</groupId>
+      <artifactId>pax-logging-api</artifactId>
+      <version>${pax.logging.jetty12.version}</version>
+    </dependency>
+    <!--
+      Pin explicitly, same reason as org.jupnp.osgi.tests.jetty12's own pom.xml: the itests parent pom
+      pulls in org.jupnp.bom.runtime (needed by org.jupnp.osgi.tests for the javax pax-web-jetty 8.0.25),
+      which transitively drags in jetty-util at the old Jetty 9.4.x version. A direct dependency
+      declaration always wins over a transitive one.
+    -->
+    <dependency>
+      <groupId>org.eclipse.jetty</groupId>
+      <artifactId>jetty-util</artifactId>
+      <version>${jetty12.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.aries.spifly</groupId>
+      <artifactId>org.apache.aries.spifly.dynamic.bundle</artifactId>
+      <version>1.3.7</version>
+    </dependency>
+  </dependencies>
+
+</project>
